third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/column-auto-repeat-012-expected.txt

This is a testharness.js-based test.
[FAIL] .masonry 1
  assert_equals: \n<div class="masonry" data-expected-width="250" data-expected-height="125">\n  <div class="item" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="125"></div>\n</div>\nwidth expected 100 but got 250
[FAIL] .masonry 2
  assert_equals: \n<div class="masonry" style="width: 200px; height: 100px;" data-expected-width="250" data-expected-height="125">\n  <div class="item" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="125"></div>\n</div>\nwidth expected 100 but got 250
[FAIL] .masonry 3
  assert_equals: \n<div class="masonry" style="width: min-content; height: min-content;" data-expected-width="250" data-expected-height="125">\n  <div class="item" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="125"></div>\n</div>\nwidth expected 100 but got 250
[FAIL] .masonry 4
  assert_equals: \n<div class="masonry" style="width: max-content; height: max-content;" data-expected-width="250" data-expected-height="125">\n  <div class="item" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="125"></div>\n</div>\nwidth expected 100 but got 250
Harness: the test ran to completion.
